refactor(proactive): split idle.py into config/scanner/installer
386-line idle.py 를 책임별 3개 파일로 쪼개고 Pydantic
설정은 top-level config/ 일관성에 맞춰 분리.

- config/idle.py: IdleConfig, QuietHours, _DEFAULT_IDLE_PROMPT
- services/proactive/scanner.py: IdleScanner + Protocols + 내부 헬퍼
- services/proactive/installer.py: install_idle_* + 상수
- services/proactive/idle.py 제거
- 테스트 import 갱신

콘텐츠 변경 없음. 213 passed 유지.
